pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "simple-rag-writer"
version = "0.0.1"
description = "Simplified rag_writer-style tool with MCP and multi-provider LLMs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.0",
  "PyYAML>=6.0",
  "rich>=13.0",
  "litellm>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "black>=24.0",
  "mypy>=1.0",
]

[project.scripts]
srw = "simple_rag_writer.cli.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
